refactor: alter package name for internal errors
