+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/capture/internal/errors/EmbraceInternalErrorService.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.logging
+package io.embrace.android.embracesdk.capture.internal.errors
 
 import io.embrace.android.embracesdk.config.ConfigService
 import io.embrace.android.embracesdk.internal.clock.Clock
